+++ lib/czech_post_b2b_client/printing_templates.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+# rubocop:disable Style/AsciiComments
+
 # Available printing templates, which can be used in AdressSheetsGenerator `options[:template_id]`
 #
 # Do not miss `CzechPostB2bClient::PrintingTemplates.all_classes` method,
@@ -12,10 +14,15 @@ module CzechPostB2bClient
       end
     end
 
-    module AdressLabel
+    module AddressLabel
       class Simple < Base
         @id = 7
         @description = 'adresní štítek (alonž) - samostatný'
+      end
+
+      class SimpleA6FourOnPage < Base
+        @id = 103
+        @description = 'Adresní štítek A6 – 4x'
       end
 
       class SimpleWithCODVoucherA < Base
@@ -23,17 +30,17 @@ module CzechPostB2bClient
         @description = 'adresní štítek (alonž) + dobírková poukázka A'
       end
 
-      class Blank4OnPage < Base
+      class BiancoFourOnPage < Base
         @id = 20
         @description = 'adresní štítek bianco - 4x (A4)'
       end
 
-      class Blank < Base
+      class Bianco < Base
         @id = 21
         @description = 'adresní štítek bianco - samostatný'
       end
 
-      class BlankLandscape < Base
+      class BiancoLandscape < Base
         @id = 39
         @description = 'adresní štítek bianco - samostatný (na šířku)'
       end
@@ -41,22 +48,22 @@ module CzechPostB2bClient
       module ForeignPackage
         class Standard < Base
           @id = 58
-          @description = 'CP72 - standardní balík do zahraničí'
+          @description = 'CP72 - standardní balík do zahraničí' # or 'AŠ - samostatný Standardní balík do zahraničí'
         end
 
-        class Standard2OnPage < Base
+        class StandardTwoOnPage < Base
           @id = 59
-          @description = 'CP72 - standardní balík do zahraničí (2x A4 )'
+          @description = 'CP72 - standardní balík do zahraničí (2x A4 )' # or 'AŠ - 4xA4 Standardní balík do zahraničí'
         end
 
-        class Valuable < Base
+        class Insured < Base
           @id = 60
-          @description = 'CP72 - cenný balík do zahraničí'
+          @description = 'CP72 - cenný balík do zahraničí' # or 'AŠ - samostatný Cenný balík do zahraničí'
         end
 
-        class Valuable4OnPage < Base
+        class InsuredFourOnPage < Base
           @id = 61
-          @description = 'CP72 - cenný balík do zahraničí (2x A4)'
+          @description = 'CP72 - cenný balík do zahraničí (2x A4)' # or '4xA4 Cenný balík do zahraničí'
         end
 
         class EMS < Base
@@ -64,21 +71,9 @@ module CzechPostB2bClient
           @description = 'AŠ - samostatný EMS zahraničí'
         end
 
-        class EMS2x4 < Base
+        class EMSTwoOnPage < Base
           @id = 63
           @description = 'AŠ - 2xA4 EMS do zahraničí'
-        end
-
-        # Formulare ID 72 a 73 je mozno pouzit pouze pro zasilky s prefixem CE
-        # do zemi AT, DE, FR, GR, HR, CH, IS, LU, LV, NO, PL, SK
-        class HarmonizedLabel < Base
-          @id = 72
-          @description = 'Obchodní balík do zahraničí – samostatný'
-        end
-
-        class HarmonizedLabel4OnPage < Base
-          @id = 73
-          @description = 'Obchodní balík do zahraničí –  4x (A4)'
         end
       end
     end
@@ -110,7 +105,7 @@ module CzechPostB2bClient
       @description = 'poštovní dobírková poukázka A - samostatná'
     end
 
-    class CODVoucherA3OnPage < Base
+    class CODVoucherAThreeOnPage < Base
       @id = 11
       @description = 'poštovní dobírková poukázka A - 3x (A4)'
     end
@@ -145,7 +140,7 @@ module CzechPostB2bClient
       @description = 'Dodejka'
     end
 
-    module CustomDeclaration
+    module CustomsDeclaration
       class CN22 < Base
         @id = 56
         @description = 'Celní prohlášení CN22 - 1x (A4)'
@@ -156,7 +151,7 @@ module CzechPostB2bClient
         @description = 'Celní prohlášení CN23 - 1x (A4)'
       end
 
-      class CN224OnPage < Base
+      class CN22FourOnPage < Base
         @id = 74
         @description = 'Celní prohlášení CN22 - 4x (A4)'
       end
@@ -166,14 +161,66 @@ module CzechPostB2bClient
         @description = 'Celní prohlášení CN22  s ČK - 1x (A4)'
       end
 
-      class CN22WithCK4OnPage < Base
+      class CN22WithCKFourOnPage < Base
         @id = 76
         @description = 'Celní prohlášení CN22  s ČK - 4x (A4)'
       end
     end
+
+    class CN22FourOnPage < Base
+      @id = 74
+      @description = 'Celní prohlášení CN22 - 4x (A4)'
+    end
+
+    module HarmonizedLabel
+      # Formulare ID 72 a 73 je mozno pouzit pouze pro zasilky s prefixem CE
+      # do zemi AT, DE, FR, GR, HR, CH, IS, LU, LV, NO, PL, SK
+      class Simple < Base
+        @id = 72
+        @description = 'Obchodní balík do zahraničí – samostatný' # or 'Harmonizovaný štítek pro MZ produkty-samostatný'
+      end
+
+      class SimpleFourOnPage < Base
+        @id = 73
+        @description = 'Obchodní balík do zahraničí –  4x (A4)' # or 'Harmonizovaný štítek pro MZ produkty – 4x (A4)'
+      end
+
+      class BiancoFourOnPage < Base
+        @id = 100
+        @description = 'Harmonizovaný štítek bianco – 4x (A4)'
+      end
+
+      class Bianco < Base
+        @id = 101
+        @description = 'Harmonizovaný štítek bianco – samostatný'
+      end
+
+      class BiancoPortrait < Base
+        @id = 102
+        @description = 'Harmonizovaný štítek bianco – samostatný (na výšku)'
+      end
+
+      class ZebraBianco105x148 < Base
+        @id = 200
+        @description = 'Harmonizovaný štítek bianco - (Zebra - 105x148)'
+      end
+
+      class ZebraBianco100x150 < Base
+        @id = 201
+        @description = 'Harmonizovaný štítek bianco - (Zebra - 100x150)'
+      end
+
+      class ZebraBianco100x125 < Base
+        @id = 202
+        @description = 'Harmonizovaný štítek bianco - (Zebra - 100x125)'
+      end
+    end
+
     # has to be at the end, to load all subcasses before
     def self.all_classes
       ObjectSpace.each_object(CzechPostB2bClient::PrintingTemplates::Base.singleton_class)
     end
   end
 end
+
+# rubocop:enable Style/AsciiComments