+++ test/czech_post_b2b_client_test.rb
@@ -14,7 +14,7 @@ class CzechPostB2bClientTest < Minitest::Test
   def test_knows_available_templates
     all_templates = CzechPostB2bClient::PrintingTemplates.all_classes.to_a
 
-    assert_equal 31, all_templates.size
+    assert_equal 39, all_templates.size
 
     template_61 = all_templates.detect { |t| t.id == 61 }
     assert_equal 'CP72 - cenný balík do zahraničí (2x A4)', template_61.description